# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventory",
        "description": "Documentação da API Inventory gerada automaticamente via Swagger.",
        "contact": {
            "email": "victor.hugo@bizsys.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stores": {
            "get": {
                "tags": [
                    "Lojas"
                ],
                "summary": "Lista lojas paginadas",
                "description": "Retorna uma lista paginada de lojas, incluindo contatos associados.",
                "operationId": "789854b8f9730d47f1ba52ca5e610238",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (padrão: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de lojas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de lojas obtida com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id_store": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Loja XPTO"
                                                            },
                                                            "contacts": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id_contact": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "name": {
                                                                            "type": "string",
                                                                            "example": "Contato 1"
                                                                        },
                                                                        "email": {
                                                                            "type": "string",
                                                                            "example": "contato@email.com"
                                                                        },
                                                                        "phone": {
                                                                            "type": "string",
                                                                            "example": "11999999999"
                                                                        },
                                                                        "observation": {
                                                                            "type": "string",
                                                                            "example": "Observação"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            },
            "post": {
                "tags": [
                    "Lojas"
                ],
                "summary": "Cria uma nova loja",
                "description": "Cria uma loja e associa contatos.",
                "operationId": "5562ee6fb25788924172167093719fe9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "contacts"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Loja XPTO"
                                    },
                                    "contacts": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "name",
                                                "email",
                                                "phone"
                                            ],
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Contato 1"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "contato@email.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "11999999999"
                                                },
                                                "observation": {
                                                    "type": "string",
                                                    "example": "Observação"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loja criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Loja criada com sucesso."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Loja XPTO"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno ao criar loja"
                    }
                }
            }
        },
...
    },
...
}